+++ CHANGELOG.md
@@ -13,12 +13,28 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
+## [15.2.0]
+
+### Changed
+
+- Exclude sub-projects in gradle-dependencies-q.txt unless keep_subprojects_in_maven_tree is true.
+- The format of Gradle sub-project GAVs that were found in a dependency tree has changed from "internal:my-subproject:1.0.0" to ":my-subproject:*".
+- Gradle dependencies in gradle-dependencies-q.txt with "(n)" suffix will be ignored in favor of their resolved dependencies, so the full dependency list could now be smaller but still correct.
+
+### Added
+
+- Added support for another project name pattern in gradle-dependencies-q.txt lockfiles.
+
 ## [15.1.3]
 
 ### Added
 
 - Start collecting "project_name" from gradle-dependencies-q.txt lockfiles.
 - Start collecting "project_name" from package.json manifests.
+
+### Changed
+
+- Bibliothecary::Parsers::Maven's GRADLE_PROJECT_REGEXP constant was renamed to GRADLE_DEPENDENCY_PROJECT_REGEXP.
 
 ### Removed